// src/interfaces/product.ts
export interface Product {
  abv: number;
  category: string;
  description: string;
  details: string;
  flag: string;
  intensity: number;
  name: string;
  origin: string;
  packshot: string;
  price: number;
  smell: string;
  taste: string;
  use: string;
  useIcons: string;
  volume: number;
}